move constantlengthdataset to "advanced dataset" to avoid confusing beginners
